+++ test/integration/default/serverspec/spec_helper.rb
@@ -1,0 +1,7 @@
+#
+# Copyright (c) 2015 Nordstrom, Inc.
+# Licensed for use with the Apache2 license
+
+require 'serverspec'
+
+set :backend, :exec